Move the original query implementation to a subdirectory.
So we can still test and benchmark, but leaves the main package clear for the
new implementation.

Also: emove unused test case fields

All the tests in this group are expected to compile.
Remove the compile-error check field, which was always false.

All the tests in this group do not want a match error.
Remove the match-error check field.